--- a/README.md
+++ b/README.md
@@ -15,3 +15,9 @@ $ PORT=8080 yarn dev # override the port to listen on
 $ yarn test
 $ yarn test path/to/specific/test.ts
 ```
+
+## Update JSON Schema
+
+Update JSON schema based on TypeScript definitions:
+
+$ yarn generate-schema