--- a/package.json
+++ b/package.json
@@ -17,7 +17,8 @@
     "upload-artifact": "node-riffraff-artifact",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
-    "cdk": "cdk"
+    "cdk": "cdk",
+    "generate-schema": "ts-node ./src/scripts/generate-json-schema.ts"
   },
   "dependencies": {
     "@guardian/src-foundations": "^0.12.4",
@@ -28,6 +29,7 @@
     "emotion": "^10.0.27",
     "emotion-server": "^10.0.27",
     "express": "^4.17.1",
+    "jsonschema": "^1.2.5",
     "node-fetch": "^2.6.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0"
@@ -72,6 +74,7 @@
     "ts-jest": "^24.3.0",
     "ts-loader": "^6.2.1",
     "ts-node-dev": "^1.0.0-pre.44",
-    "typescript": "^3.7.4"
+    "typescript": "^3.7.4",
+    "typescript-json-schema": "^0.42.0"
   }
 }