--- a/src/components/ContributionsEpic.testData.ts
+++ b/src/components/ContributionsEpic.testData.ts
@@ -1,4 +1,4 @@
-import { EpicTracking, EpicLocalisation, EpicTargeting } from './ContributionsEpic';
+import { EpicTracking, EpicLocalisation, EpicTargeting } from './ContributionsEpicTypes';
 
 const content = {
     heading: 'Since you’re here...',
@@ -38,30 +38,22 @@ const targeting: EpicTargeting = {
         {
             id: 'culture/david-schwimmer',
             type: 'Keyword',
-            title: 'David Schwimmer',
         },
         {
             id: 'tv-and-radio/friends',
             type: 'Keyword',
-            title: 'Friends',
         },
         {
             id: 'tone/interview',
             type: 'Tone',
-            title: 'Interviews',
         },
         {
             id: 'publication/theguardian',
             type: 'Publication',
-            title: 'The Guardian',
         },
         {
             id: 'profile/davidsmith',
             type: 'Contributor',
-            title: 'David Smith',
-            twitterHandle: 'smithinamerica',
-            bylineImageUrl:
-                'https://i.guim.co.uk/img/uploads/2017/10/06/David-Smith,-L.png?width=300&quality=85&auto=format&fit=max&s=9aebe85c96f6f72a6ba6239cdfaed7ec',
         },
     ],
 };