This PR introduces a JSON schema document to illustrate the format that the `POST /epic` endpoint expects of the request body.

We also now use the schema to validate the incoming payload. If the request payload isn't the correct format then we'll return `400 Bad Request`.

Anecdotally the validation seems to increase response times by a few milliseconds, but this doesn't seem problematic.

Since this adds even more logic to `src/server.tsx` We have a refactoring to split that file up into separate concerns on another branch (but didn't want to include on this PR because it would have obscured the json schema changes).