@tjmw 2020-02-11
Thanks for the suggestions @nicl! I pushed b8996a4 which adds a command to generate JSON schema from the TypeScript types, instead of duplicating. Right now you always have to run manually. I'll explore some options for how to remove that friction from the workflow (with a pre-commit hook or something). There's another commit in here to refactor the types out to another file since I couldn't get `typescript-json-schema` working with tsx files.