@tjmw 2020-02-11
## Benchmarking results (Valid payload)

### Without JSON schema validation (before)

```
$ echo "POST http://localhost:8080/epic" | vegeta attack -rate=10/s -duration=60s -body=payload.json -header="Content-Type: application/json" > resultsWithoutJsonSchema.gob
$ vegeta report resultsWithoutJsonSchema.gob
```

```
Requests      [total, rate, throughput]  600, 10.02, 10.02
Duration      [total, attack, wait]      59.902481643s, 59.899925884s, 2.555759ms
Latencies     [mean, 50, 95, 99, max]    2.819333ms, 2.782994ms, 3.923562ms, 5.577767ms, 7.740792ms
Bytes In      [total, mean]              2913000, 4855.00
Bytes Out     [total, mean]              859200, 1432.00
Success       [ratio]                    100.00%
Status Codes  [code:count]               200:600
Error Set:
```

### With JSON schema validation (after)

```
$ echo "POST http://localhost:8080/epic" | vegeta attack -rate=10/s -duration=60s -body=payload.json -header="Content-Type: application/json" > resultsWithJsonSchema.gob
$ vegeta report resultsWithJsonSchema.gob
```

```
Requests      [total, rate, throughput]  600, 10.02, 10.02
Duration      [total, attack, wait]      59.909900569s, 59.906700536s, 3.200033ms
Latencies     [mean, 50, 95, 99, max]    3.095149ms, 3.186452ms, 4.548159ms, 6.708991ms, 8.404663ms
Bytes In      [total, mean]              2913000, 4855.00
Bytes Out     [total, mean]              859200, 1432.00
Success       [ratio]                    100.00%
Status Codes  [code:count]               200:600
Error Set:
```