@andre1050 2020-02-12
There's a follow up PR to this work here, to run the regenerate task on `pre-push` using Husky.
https://github.com/guardian/contributions-service/pull/35