@tjmw 2020-02-13
@nicl I increased the rate to 50/second and re-ran the test against CODE.

## Without JSON Schema Validation

```
$ echo "POST https://u9q10wj8a0.execute-api.eu-west-1.amazonaws.com/epic" | vegeta attack -rate=50/s -duration=60s -body=payload.json -header="Content-Type: application/json" > resultsCodeWithoutJsonSchema.gob
$ vegeta report resultsCodeWithoutJsonSchema.gob

Requests      [total, rate, throughput]  3000, 50.01, 49.96
Duration      [total, attack, wait]      1m0.044588198s, 59.983088076s, 61.500122ms
Latencies     [mean, 50, 95, 99, max]    89.567336ms, 62.424217ms, 125.452023ms, 1.293479757s, 1.580886292s
Bytes In      [total, mean]              14565000, 4855.00
Bytes Out     [total, mean]              4296000, 1432.00
Success       [ratio]                    100.00%
Status Codes  [code:count]               200:3000
Error Set:
```

## With JSON Schema Validation

```
$ echo "POST https://u9q10wj8a0.execute-api.eu-west-1.amazonaws.com/epic" | vegeta attack -rate=50/s -duration=60s -body=payload.json -header="Content-Type: application/json" > resultsCodeWithJsonSchema.gob
$ vegeta report resultsCodeWithJsonSchema.gob

Requests      [total, rate, throughput]  3000, 50.02, 49.95
Duration      [total, attack, wait]      1m0.05802208s, 59.976723847s, 81.298233ms
Latencies     [mean, 50, 95, 99, max]    106.39074ms, 63.97847ms, 166.845006ms, 1.346176813s, 1.604533614s
Bytes In      [total, mean]              14565000, 4855.00
Bytes Out     [total, mean]              4296000, 1432.00
Success       [ratio]                    100.00%
Status Codes  [code:count]               200:3000
Error Set:
```